added  betetr method for getting all job on pages

+++ jobApp/jobEngine/jobParserLinkedin.py
@@ -3,6 +3,10 @@ from bs4 import BeautifulSoup, Comment
 from seleniumWrapper import WebScraper
 import json
 import csv
+import os
+import re
+from linkedinEasyApplyLegacyCode import EasyApplyLinkedin
+
 class JobParser:
     def __init__(self, linkedin_data):
         """Parameter initialization"""
@@ -25,6 +29,9 @@ class JobParser:
         self.easyApplyList = []
         self.offsiteApplyList = []
         self.officialJobLinks = []
+        # the bot
+        self.bot = EasyApplyLinkedin('jobApp/secrets/linkedin.json', headless=True)
+
 
     def setEasyApplyFilter(self, easy_apply_filter=False):
         print("Warning: easy apply filter is only visible for logged user ")
@@ -32,8 +39,18 @@ class JobParser:
         # set the filter to true or false
         self.params['f_AL']=easy_apply_filter
 
+    def generateLinksPerPage(self, max_pages = 5)-> list: #125 jobs
+        for _ in range(max_pages):
+            self.generateLinks()
+        #@NOTE: added save to csv to keep track of links
+        if self.filter_easy_apply:
+            self.saveLinksToCsv(links= [self.easyApplyList, "na"])
+            return [self.easyApplyList, "na"]
+        else:
+            self.saveLinksToCsv(links= [self.offsiteApplyList, self.officialJobLinks ])
+            return [self.offsiteApplyList, self.officialJobLinks ]
+        
     def generateLinks(self):
-
         full_url = f"{self.base_url}?{'&'.join([f'{k}={v}' for k, v in self.params.items()])}"
         print(f"constructed url: {full_url }")
         response = requests.get(self.base_url, params=self.params)
@@ -62,19 +79,34 @@ class JobParser:
                     links.append(href)  # keep track of global links
                 else:
                     self.filterJobList(href, False, True)
-                    links.append(href)      
-
-        print(f"total links found {len(links)}")
+                    links.append(href)              
         self.jobList+= links
+        print(f"total links found {len(self.jobList)}")
         if self.filter_easy_apply:
             print(f"total links easy apply found {len(self.easyApplyList)}")  
             return  [self.easyApplyList, "na"]
         else:
             print(f"total links offsite apply found {len(self.offsiteApplyList)}") 
             return [self.offsiteApplyList, self.officialJobLinks ]
-    
-    def filterJobList(self, job_href, onsite=False, offsite=False )-> list:
         
+    def generateLinksSeleniumV2(self, pages=5):
+        self.bot.getEasyApplyJobSearchUrlResults()
+        self.jobList = self.bot.getUnlockJobLinksNoLogin(pages) # we have all links
+        for link in self.jobList:
+            if self.filter_easy_apply: # if we are looking only for easy apply jobs
+                self.filterJobList(link, True, False)
+            else:
+                self.filterJobList(link, False, True)
+        if self.filter_easy_apply:
+            print(f"total links easy apply found {len(self.easyApplyList)}")  
+            self.saveLinksToCsv(links= [self.easyApplyList, "na"])
+            return  [self.easyApplyList, "na"]
+        else:
+            print(f"total links offsite apply found {len(self.offsiteApplyList)}") 
+            self.saveLinksToCsv(links= [self.offsiteApplyList, self.officialJobLinks ])
+            return [self.offsiteApplyList, self.officialJobLinks ]
+
+    def filterJobList(self, job_href, onsite=False, offsite=False )-> list:
             response = requests.get(job_href)
             html_source = response.content
             # Create a BeautifulSoup object to parse the HTML source code
@@ -108,39 +140,51 @@ class JobParser:
         # print the text content of the comment node
         print(f"official job link: {official_link}\n")
         self.officialJobLinks.append(official_link)
-        
-    def generateLinksPerPage(self, max_pages = 5)-> list: #125 jobs
-        for _ in range(max_pages):
-            self.generateLinks()
-        #@NOTE: added save to csv to keep track of links
-        if self.filter_easy_apply:
-            self.saveLinksToCsv(links= [self.easyApplyList, "na"])
-            return [self.easyApplyList, "na"]
-        else:
-            self.saveLinksToCsv(links= [self.offsiteApplyList, self.officialJobLinks ])
-            return [self.offsiteApplyList, self.officialJobLinks ]  
-    
-    def saveLinksToCsv(self, links,  csv_file='jobApp/data/links.csv'):
-        # Open the CSV file in write mode
-        with open(csv_file, mode='w', newline='') as file:
-            writer = csv.writer(file)
-            # Write the header row
-            writer.writerow(['id', 'keyword', 'location', 'internal link', 'external link'])
-            # Write each row of data to a new row in the CSV file
-            for i, link in enumerate(links[0]):
-                writer.writerow([i+1, self.job_title, self.location, link, links[1][i] ])
-        print(f"links saved to {csv_file}")
+         
+    def extract_link_id(self, link:str):
+        # find the first occurrence of "?"
+        question_mark_index = link.find("?")
+        # extract the substring before "?"
+        id_string = link[:question_mark_index]
+        # find the last occurrence of "-" before "?"
+        dash_index = id_string.rfind("-")
+        # extract the substring after the last "-"
+        id = id_string[dash_index+1:]
+        return id
 
-    def _use_selenium_to_get_easy_apply_jobs(self):
-        scraper = WebScraper('jobApp/secrets/linkedin.json', headless=False)
-        scraper.bot.login_linkedin()
-        scraper.bot.getEasyApplyJobSearchUrlResults(jobParserObj.base_url, jobParserObj.params)
-        linksToApply = scraper.bot.getJobOffersListEasyApply()
+    def saveLinksToCsv(self, links, csv_file='jobApp/data/links.csv'):
+        # Check if the CSV file exists: read and append only new links
+        ids = list()
+        counter = 0
+        if os.path.isfile(csv_file):
+            # Read
+            with open(csv_file, mode='r', newline='') as file:
+                reader = csv.reader(file)
+                next(reader)  # Skip header row
+                for row in reader:
+                    ids.append(row[3]) # we get all ids there
+            # write
+            
+            with open(csv_file, mode='a', newline='') as file:
+                writer = csv.writer(file)
+                for i, link in enumerate(links[0]): # new links loop
+                    if self.extract_link_id(link) not in ids: 
+                        counter +=1
+                        writer.writerow([len(ids)+counter, self.job_title, self.location, self.extract_link_id(link) , link, links[1][i]])
+        # no csv, write new from zero
+        else: 
+            with open(csv_file, mode='w', newline='') as file:
+                writer = csv.writer(file)
+                # Write the header row if the file is empty
+                if os.stat(csv_file).st_size == 0:
+                    writer.writerow(['id', 'keyword', 'location','link id', 'internal link', 'external link'])
+                for i, link in enumerate(links[0]):
+                    writer.writerow([i+1, self.job_title, self.location, self.extract_link_id(link) , link, links[1][i]])
+        print(f"Links saved to {csv_file}")
 
-        return linksToApply
 
 if __name__ == '__main__':
 
    jobParserObj= JobParser('jobApp/secrets/linkedin.json')
    jobParserObj.setEasyApplyFilter(False)
-   jobsLinks = jobParserObj.generateLinksPerPage(1)
+   jobsLinks = jobParserObj.generateLinksSeleniumV2()
